footy-probability-app/requirements.txt
streamlit>=1.36
pandas>=2.2
numpy>=1.26
python-dotenv>=1.0
requests>=2.31
scikit-learn>=1.4
